E02: Updated db script for voluntary_work for voluntary work credit project. - changed integer field to float
git-svn-id: https://svn.csc.ncsu.edu:8443/svn/research-expertiza/projectA/trunk@1132 392d7b7b-3f31-0410-9dc3-c95a9635ea79

+++ db/migrate/20101120220132_create_voluntary_works.rb
@@ -2,7 +2,7 @@ class CreateVoluntaryWorks < ActiveRecord::Migration
   def self.up
     create_table :voluntary_works do |t|
       t.string :name
-      t.int :weight
+      t.float :weight
       t.string :type
 
       t.timestamps
